Add JsonCoder as default Coder #81
