+++ __tests__/ovm/deciders/NotDecider.test.ts
@@ -1,6 +1,6 @@
 import { Property } from '../../../src/ovm/types'
 import { Bytes } from '../../../src/types/Codables'
-import EthCoder from '../../../src/coder/EthCoder'
+import Coder from '../../../src/coder'
 import {
   initializeDeciderManager,
   NotDeciderAddress,
@@ -15,17 +15,13 @@ describe('NotDecider', () => {
   const deciderManager = initializeDeciderManager()
   it('decide not(false)', async () => {
     const decision = await deciderManager.decide(
-      new Property(NotDeciderAddress, [
-        EthCoder.encode(falseProperty.toStruct())
-      ])
+      new Property(NotDeciderAddress, [Coder.encode(falseProperty.toStruct())])
     )
     expect(decision.outcome).toEqual(true)
   })
   it('decide not(true)', async () => {
     const decision = await deciderManager.decide(
-      new Property(NotDeciderAddress, [
-        EthCoder.encode(trueProperty.toStruct())
-      ])
+      new Property(NotDeciderAddress, [Coder.encode(trueProperty.toStruct())])
     )
     expect(decision.outcome).toEqual(false)
     // valid challenge is SampleDecider(true)
