+++ src/coder/index.ts
@@ -1,4 +1,4 @@
 export * from './Coder'
 export * from './Error'
-import EthCoder from './EthCoder'
-export default EthCoder
+import JsonCoder from './JsonCoder'
+export default JsonCoder
